fix: swap box context providers
