+++ packages/app/src/MainUI.tsx
@@ -1,15 +1,21 @@
-import { Suspense, lazy } from 'react'
+import { Suspense, lazy, useEffect } from 'react'
 import { HashRouter, Navigate, Route, Routes } from 'react-router-dom'
+import { SNSAdaptorContextRef } from '@masknet/plugin-infra/content-script'
 import { DashboardForDesktop } from './components/DashboardDesktop.js'
 import { DashboardForMobile } from './components/DashboardMobile.js'
 import { DashboardContext } from './contexts/DashboardContext.js'
 import { ApplicationRoutes } from './constants/ApplicationRoutes.js'
+import { createSharedContext } from './helpers/createSharedContext.js'
 
 const OverviewPage = lazy(() => import(/* webpackPrefetch: true */ './pages/OverviewPage.js'))
 const ExplorerPage = lazy(() => import(/* webpackPrefetch: true */ './pages/ExplorerPage.js'))
 const SwapPage = lazy(() => import(/* webpackPrefetch: true */ './pages/SwapPage.js'))
 
 export function MainUI() {
+    useEffect(() => {
+        SNSAdaptorContextRef.value = createSharedContext()
+    }, [])
+
     return (
         <DashboardContext.Provider>
             <Suspense fallback={null}>
